+++ packages/amplify-e2e-core/src/utils/feature-flags.ts
@@ -1,6 +1,5 @@
 import _ from 'lodash';
-import { JSONUtilities } from '@aws-amplify/amplify-cli-core';
-import { pathManager, FeatureFlagsEntry } from '@aws-amplify/amplify-cli-core';
+import { JSONUtilities, pathManager, FeatureFlagsEntry } from '@aws-amplify/amplify-cli-core';
 
 type FeatureFlagData = { features: FeatureFlagsEntry };
 const getFeatureFlagFilePath = (projectRoot: string) => {