+++ packages/graphql-predictions-transformer/src/resources.ts
@@ -19,8 +19,7 @@ import {
   HttpMappingTemplate,
 } from 'graphql-mapping-template';
 import { IAM, Fn, AppSync, Lambda } from 'cloudform-types';
-import { HttpConfig, LambdaConfig } from 'cloudform-types/types/appSync/dataSource';
-import DataSource from 'cloudform-types/types/appSync/dataSource';
+import DataSource, { HttpConfig, LambdaConfig } from 'cloudform-types/types/appSync/dataSource';
 import { Policy } from 'cloudform-types/types/iam/group';
 import { iamActions } from './predictions_utils';