+++ packages/graphql-transformers-e2e-tests/src/__tests__/NonModelAuthFunction.e2e.test.ts
@@ -445,7 +445,7 @@ afterAll(async () => {
 /**
  * Test queries below
  */
-test('Test calling echo function as a user via IAM', async () => {
+test('calling echo function as a user via IAM', async () => {
   const query = gql`
     query {
       echo(msg: "Hello")