+++ packages/amplify-graphql-schema-generator/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.4.1...@aws-amplify/graphql-schema-generator@0.4.2) (2023-08-30)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-generator
+
+## [0.4.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.4.0...@aws-amplify/graphql-schema-generator@0.4.1) (2023-08-28)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-generator
+
 # [0.4.0](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-generator@0.3.0...@aws-amplify/graphql-schema-generator@0.4.0) (2023-08-09)
 
 ### Features